requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-chess==1.999
numpy==1.26.4
torch==2.3.1
scikit-learn==1.5.1
tqdm==4.66.4
pandas==2.2.2
requests==2.32.3
